pyproject.toml
[project]
name = "agro-api"
version = "0.2.2"
description = "API that allows owners to manage their agro estates"
authors = [
    {name = "Fernando Caçador",email = "fernando.cacador@proton.me"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "fastapi[standard] (>=0.127.1,<0.128.0)",
    "sqlalchemy[asyncio] (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "fastapi-injectable (>=1.4.6,<2.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "faker (>=40.1.0,<41.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['E402', 'F811']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
run = 'fastapi dev agro_api/app.py --host 0.0.0.0'
serve = 'uvicorn agro_api.app:app --workers 2'

lint = 'ruff check'

pre_format = 'ruff check --fix'
format = 'ruff format'

pre_test = 'task lint'
test = 'pytest -s -x --cov=agro_api -vv'
post_test = 'coverage html'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]
